backend/requirements.txt
flask
flask-mysqldb
pyjwt
werkzeug
Flask-Cors